feat: modified code structure and validate script
