+++ src/cofi_espresso/__init__.py
@@ -1,0 +1,17 @@
+from .espresso_problem import EspressoProblem
+
+
+try:
+    from . import _version
+
+    __version__ = _version.__version__
+except ImportError:
+    pass
+
+
+__all__ = [
+    "EspressoProblem",
+]
+
+# from .example_name import ExampleName
+# __all__.append("ExampleName")
