--- utils/build_package/_package_src/src/cofi_espresso/__init__.py
@@ -1,6 +1,0 @@
-try:
-    from . import _version
-
-    __version__ = _version.__version__
-except ImportError:
-    pass
